fix: fix structure page and modularity feat: add alert dialog delete position and department
